Create routes module
Split route source file into different source files.
- federated
- network
- general
