+++ app/main/node_routing/__init__.py
@@ -1,0 +1,3 @@
+from .nodes_network import NetworkManager
+
+network_manager = NetworkManager()
